requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
sqlalchemy==2.0.27
google-generativeai==0.4.1
python-dotenv==1.0.1
pydantic==2.6.1
